sum(): associated-Output PySum with generic bounds and use-context pinning (issue #133)
The sum() builtin was broken in both directions: the PySum<T> trait was
implemented only for slice types, so even a concrete `sum(xs)` on a
Vec<i64> local failed at rustc, and a generic parameter got no bound at
all (the issue's calc repro). Redesign, completing #133's deferred
piece:

- stdpython: PySum carries its OUTPUT as an associated type — every
  iterable sums to exactly one result type, so bounds compose with the
  operator-Output machinery and a use context can pin the result.
  Impls for owned/borrowed Vec and slices of i64 and f64 (i32/f32 are
  dropped: rython's numeric model is i64/f64, and an i32 impl would
  leave `sum(vec![1, 2, 3])` ambiguous over {integer}, with rustc's
  i32 literal fallback contradicting an i64 use context), a bool list
  counting the Trues (bool ⊂ int), and PyList.
- param_infer.rs: `sum(p)` on an unannotated parameter emits
  `T: PySum` (builtin table + M5 satisfiability: numeric/bool lists
  satisfy, a str/dict argument is CPython's TypeError, loud);
  `return sum(p)` projects `<T as PySum>::Output`, so ONE generic
  function serves int and float lists.
- param_infer.rs: a sum() result stored into an ALREADY-TYPED slot pins
  the Output — `chunks = [sum(xs)]` after the i64-seeded
  `chunks.append(len(xs))` emits `T: PySum<Output = i64>` (subsuming
  the plain bound), which rustc will not infer across the generic
  boundary. The per-body type analysis supplies the slot's final type,
  the same source the empty-list store renders from.
- Two hand-written runtime tests that pinned the removed i32 impls
  update to the i64 model the compiler actually generates.

The issue's calc repro now runs end to end: pinned against python3 at
codegen (codegen_semantics.rs) and runtime (convert_tests.rs:
[6] / 6 / 4.0 / 2 / 60 — calc, generic int and float sums, a bool
list, and a concrete literal list). spec §10.1 records the design.

+++ crates/stdpython/tests/integration_comprehensive.rs
@@ -271,12 +271,14 @@ fn test_complex_python_simulation() {
     println!("📋 Testing complex Python code simulation...");
     
     // Simulate: def calculate_stats(numbers): return {"sum": sum(numbers), "min": min(numbers), "max": max(numbers), "count": len(numbers)}
-    fn calculate_stats(numbers: &Vec<i32>) -> PyDictionary<String, i32> {
+    // Python ints are i64 in generated code (PySum is i64/f64-only —
+    // an i32 impl would leave integer-literal lists ambiguous, #133).
+    fn calculate_stats(numbers: &Vec<i64>) -> PyDictionary<String, i64> {
         let mut stats = PyDictionary::new();
         stats.set("sum".to_string(), sum(&numbers[..]));
         stats.set("min".to_string(), min(numbers).unwrap_or(0));
         stats.set("max".to_string(), max(numbers).unwrap_or(0));
-        stats.set("count".to_string(), len(numbers) as i32);
+        stats.set("count".to_string(), len(numbers) as i64);
         stats
     }
